Command refactored,New packages in tests/ file txt deleted
