+++ src/main/java/com/softserve/logstat/service/collector/CollectorFind.java
@@ -3,9 +3,8 @@ package com.softserve.logstat.service.collector;
 
 import com.softserve.logstat.model.Log;
 import com.softserve.logstat.model.report.Report;
-import com.softserve.logstat.service.Command;
+import com.softserve.logstat.model.Command;
 
-import java.util.Map;
 import java.util.stream.Stream;
 
 /**
