--- src/test/java/DELETE_ME.txt
@@ -1,1 +1,0 @@
-Delete this file(
