+++ src/test/java/logstat/service/collector/CollectorFactoryTest.java
@@ -1,5 +1,6 @@
-package com.softserve.logstat.service.collector;
+package java.logstat.service.collector;
 
+import com.softserve.logstat.service.collector.*;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -10,22 +11,22 @@ class CollectorFactoryTest {
 
     @Test
     void choose1() {
-        Collector collector1=CollectorFactory.choose("top");
+        Collector collector1= CollectorFactory.choose("top");
 
-        assertEquals(collector1.getClass(),CollectorTop.class);
+        assertEquals(collector1.getClass(), CollectorTop.class);
     }
     @Test
     void choose2() {
 
         Collector collector1=CollectorFactory.choose("stat");
 
-        assertEquals(collector1.getClass(),CollectorStat.class);
+        assertEquals(collector1.getClass(), CollectorStat.class);
     }
     @Test
     void choose3() {
         Collector collector1=CollectorFactory.choose("find");
 
-        assertEquals(collector1.getClass(),CollectorFind.class);
+        assertEquals(collector1.getClass(), CollectorFind.class);
     }
     @Test
     void choose4() {
